refactor(hooks): extract announcement formatting into hook_announcement module
`src/commands/hooks.rs` bundled three concerns: config loading + step graph
building, hook execution / announcer machinery, and announcement formatting.
The formatter is a self-contained unit shared with alias dispatch — splitting
it out clarifies hooks.rs as "config -> steps -> execution".

Move `SourcedStep`, `format_pipeline_summary`, `format_pipeline_summary_from_names`,
and `step_names_from_config` (plus their snapshot tests and the announcement-format
spec docstring) into a new `hook_announcement` module. Update `alias.rs` to import
the shared helpers from there. No behavior change.
